Moved buttons to center.
