+++ client/src/lib/assets/logo animation.gif
