+++ client/src/lib/components/fragments/Navbar/Navbar.svelte
@@ -331,7 +331,6 @@
 	import logo from "$lib/assets/logo.png";
 	import { onMount } from "svelte";
 	import { page } from "$app/stores";
-	import Button from "$lib/components/ui/Button/Button.svelte";
 	let overlay: HTMLElement;
 	let addClass = false;
 	let links: HTMLElement;
@@ -375,22 +374,12 @@
 			addClass = false;
 		}
 	}
-	const gotoRegister = () => {
-		// open/goto register page.
-	}
-	const gotoLogin = () => {
-		// open/goto login page.
-	}
 </script>
 
 <section bind:this={footerSec}>
 	<a href="/" class="logo">
 		<img src={logo} alt="Apratim Logo">
-  </a>
-	<div class="cta">
-		<Button variant="secondary" label="Login" onClick={gotoLogin} size="large" ></Button>
-		<Button variant="primary" label="Register" onClick={gotoRegister} size="large" ></Button>
-	</div>
+  	</a>
 	<div class="hamburger {addClass ? 'close' : ''}" on:keypress={toggleNavBar} on:click={toggleNavBar}>
 		<div>
 			<span class="transition1"></span>
